Make migrant executable available in project root
Changes the method a bit for locating the composer autoloader. Upon post-install or post-update, creates a script that requires vendor/bin/migrant so it is available from the root, then if possible, makes it executable. This will allow us to run migrant from the project root rather than needing to call it from vendor/bin/migrant